thief_client/requirements.txt
pygame>=2.5.0
pyserial>=3.5
requests>=2.31.0
